Implement auth using Gates

+++ app/Http/Controllers/QuestionsController.php
@@ -66,6 +66,10 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+        if (\Gate::denies('update-question',$question)){
+            abort(403, 'Access dined');
+        }
+        
         return view("questions.edit", compact("question"));
     }
 
@@ -78,6 +82,10 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if (\Gate::denies('update-question',$question)){
+            abort(403, 'Access dined');
+        }
+        
         $question->update($request->only('title','body'));
         return redirect('/question')->with('success', 'Your question has been updated');
     }
@@ -90,8 +98,10 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        if (\Gate::denies('delete-question',$question)){
+            abort(403, 'Access dined');
+        }
         $question->delete();
-
         return redirect('/question')->with('success', "Your question has been deleted");
     }
 }
